refactor(web): rename PascalCase component files to kebab-case
Tightens .ls-lint.yml to require kebab-case for all .tsx (drops
the PascalCase alternative), then renames every React component
file to match. Component identifiers stay PascalCase; only the
filenames change. One uniform rule across the codebase removes
the components-are-special carve-out and the per-directory
override that would have been needed for shadcn/ui.

Refs #162
