+++ apps/web/src/components/character-summary.tsx
