+++ apps/web/src/components/footer.tsx
