+++ apps/web/src/components/layout.tsx
@@ -3,9 +3,9 @@
 
 import { Outlet } from 'react-router-dom';
 
-import { Footer } from '@/components/Footer';
-import { Header } from '@/components/Header';
-import { Nav } from '@/components/Nav';
+import { Footer } from '@/components/footer';
+import { Header } from '@/components/header';
+import { Nav } from '@/components/nav';
 
 export function Layout() {
   return (
