+++ apps/web/src/components/nav.test.tsx
@@ -5,7 +5,7 @@ import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { describe, expect, it } from 'vitest';
 
-import { Nav } from './Nav';
+import { Nav } from './nav';
 
 function renderNav(initialRoute = '/') {
   return render(
