+++ apps/web/src/components/weapon-summary.tsx
