+++ apps/web/src/features/collection/weapons/weapon-filters.tsx
