+++ apps/web/src/features/collection/weapons/weapon-instance-sidebar.tsx
