+++ apps/web/src/main.tsx
@@ -3,7 +3,7 @@
 
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import { App } from './App.tsx';
+import { App } from './app.tsx';
 import './index.css';
 
 const rootElement = document.getElementById('root');
